Сделал 4 ДЗ переместил conftest.py в директорию с тестами, добавил два теста
